feat: national examination, student self-access, emailed password reset
The Grade 9 national examination (Diplôme) arrives as its own module rather
than as a flag on `exams`, because the paper is set and marked by the Ministry:
the school registers candidates and stores what is published. A published
result is immutable — a correction is an amendment that keeps the superseded
value and is audit-logged — and national results never enter the weighted term
average. Only a student actively enrolled in an exit-grade class may be
registered, read from `grade_levels.is_exit_grade` rather than a hard-coded
grade code.

Students reach their own record through `/students/me/...`, resolved from the
token, so no path id can be edited into someone else's file. Handing in
homework is the single write the role carries, now gated by its own
`assignments.submit` permission.

Password reset moves from a mailed link to a six digit code. That code is
guessable in a way a 32 byte token was not, so a request is burned after five
wrong entries, the reply is identical whether or not the address is registered,
and the code is returned in the response only when no mail server accepted it.

Also here:
- Grade 7-9 with the twelve-subject MoEYS curriculum, grouped and weighted
- Archived teachers and guardians can be listed and restored, matching students
- Seeds that generate a whole school: 267 students, its timetable, register,
  assessments and grades
- The API test suite now runs. It never had: the harness read the token from
  the wrong place, and suites closed the shared pool while later blocks still
  needed it, so all 55 tests had been skipping silently.

Fixes: announcement creation failed outright because one parameter served both
`created_by` and `published_by` and PostgreSQL deduced two types for it; the
archived filter returned 500 whenever it was the only filter, since an empty
condition list produced a bare `WHERE`.

+++ src/modules/announcements/announcement.repository.ts
@@ -143,8 +143,12 @@ export const insertAnnouncement = async (
         END,
         COALESCE($9, FALSE), $8::timestamptz,
         CASE WHEN $7::boolean THEN NOW() ELSE NULL END,
-        $10::timestamptz, $11, $12,
-        CASE WHEN $7::boolean THEN $12 ELSE NULL END
+        $10::timestamptz, $11, $12::bigint,
+        -- $12 is the author. It fills created_by, and published_by as well when
+        -- the announcement goes out immediately. Both uses must name the type:
+        -- without the cast PostgreSQL deduces one type from the column and
+        -- another from the CASE branch and refuses the statement.
+        CASE WHEN $7::boolean THEN $12::bigint ELSE NULL END
      )
      RETURNING id`,
     [
